net-nds/389-ds-base: Version bump to 1.3.4.8
Deps on db4.8 specifically, being the current stable, instead of
trying to determine best installed version <6 at configure time, as
discussed with various devs on irc.

Additionally, 1.3.4.7 is vuln to CVE-2016-0741 and will be removed
in the next commit.  Upstream has advised that they will not be
providing a patch against 1.3.4.7 and request users to upgrade
to 1.3.4.8

Acked-by: wibrown@redhat.com

Package-Manager: portage-2.2.27