dev-libs/wayland: Fix multilib and add missing libxml2 dependency
Multilib support was broken in previous versions, as the configure
options were in src_configure, not multilib_src_configure, and so
it always built with configuration for "native" ABI, meaning that
USE=doc with multilib was probably broken. Now with 1.10 this
becomes more apparent with a new libxml2 dependency in wayland-scanner
for DTD validation.
As we do not currently install /usr/bin/wayland-scanner for other ABIs,
just disable the dtd validation support there. Ideally we wouldn't
build it at all then, but not a clean way to do so upstream right now.
For --{dis,en}able-dtd-validation to work, we need to patch the
build system to fix the support for the configure flag.
Thanks to this we are able to add the missing libxml2 dependency
without requiring multilib libxml2 for that (and commonly then also
multilib icu), which is nice, as many users haven't had a need for
a 32bit version of those as of yet and icu is a bit bigger compile.

Remove USE=static-libs handling, as autotools-utils takes care
of it for us when it's in IUSE. Also make use of the
multilib_native_use_enable helper to simplify the ebuild.

Gentoo-bug: 575212