dev-tex/minted: fix compilation of documentation
Compiling of documentation currently failed, as -shell-escape was not
passed to pdflatex. With recent change in latex-package.eclass
(f6082c5) it is now possible to add this flag and therefore compile the
documentation.
Moreover in minted-2.1 there is an extra curly brace in minted.dtx which
leads to a compile failure also with -shell-escape specified. Until it
is fixed upstream, I added a patch to remove this extra curly brace.

Closes: https://github.com/gentoo/gentoo/pull/986
Acked-by: Alexis Ballier <aballier@gentoo.org>
Signed-off-by: Patrice Clement <monsieurp@gentoo.org>