app-text/docbook-xsl-stylesheets: Fix #575818 (Java), #536600 (Ruby)
Delete the unnecessary Java-related stuff and other tools as they
bloat the stage3 tarballs massively. Fixes bug #575818. A java USE
flag could be added but it's probably not worth the hassle.

Use the ruby-single eclass as suggested by graaff in bug #536600. I
was going to do dev-lang/ruby:* as suggested by pacho but ruby-single
seems like the correct thing to do and I don't want to go against the
Ruby team.

This also removes a useless find call.

Package-Manager: portage-2.2.28