sys-fs/lvm2: make depend() in lvm init script quiet
The depend() function makes a call to lvm which can produce output on
stderr which might seem concerning to some but is really just harmless.
We are checking to see if we need lvmetad to be able to start up properly.
At the same time make sure all possible versions receive updated init
scripts.

Gentoo-Bug: 577308
Gentoo-Bug: 573384

Package-Manager: portage-2.2.26
Signed-off-by: Doug Goldstein <cardoe@gentoo.org>