dev-util/buildbot: Added migration guide and updated setup instructions.
buildbot-0.8.12-r2 prints ewarn messages when migrating from a previous
version. The messages describe the necessary steps to migrate properly. In
addition to this, the user is notified of the support for emerge --config and
DOC_CONTENT was updated to point to the possibility of using it.

This string of 5 commits merges gentoo/gentoo#1027 and solves bug #570666.

Gentoo-Bug: https://bugs.gentoo.org/570666
Acked-By: Markos Chandras hwoarang@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/1027

Signed-off-by: Patrice Clement <monsieurp@gentoo.org>