sci-misc/boinc: New ebuild for boinc-7.2.42 with vsyscall check
Gentoo-Bug: 577490, 578750

The ebuild is essentially the same as for the new boinc-7.6.31 with
the following exception:

The 7.2 series can not be built with wxGTK:3.0, therefore this ebuild
uses wxGTK:2.8.

Further a check against VSYSCALL kernel options has been added.
Russell Knighton found out, that some projects segfault if VSYSCALL
mappings are completely disabled, and reported the issue in bug
578750.

The check against CONFIG_LEGACY_VSYSCALL_NONE and
CONFIG_X86_VSYSCALL_EMULATION is based on his patch submitted in the
bug reported by him.

Package-Manager: portage-2.2.28