dev-python/docker-py: add version 1.8.0-r1
Due to bug #580388, I've removed python2_7 support from docker-py.  This
shouldn't be a problem.  This removes the conflict due to py2-ipaddress and
makes the question of which package to keep moot.  With py2-ipaddress no longer
required by anything in the tree I will be running it through the last-rites
process.

Package-Manager: portage-2.2.28