net-analyzer/nagios-core: new patch and revision to fix logo installation.
The fix for bug 388321 required a patch, but that patch broke the
installation of the theme logos (which live in their own
directory). That patch has been updated as a new revision,
use-INSTALL-to-install-themes-r1.patch, to avoid impacting
nagios-core-4.0.8-r2.ebuild which is stable on amd64. The new revision
nagios-4.1.1-r1 should pick up the new patch.

Gentoo-Bug: 388321
Gentoo-Bug: 581148

Package-Manager: portage-2.2.26