media-video/makemkv: Launch executables from the PATH
Upstream launches wget as /usr/bin/wget and uses a hardcoded set of
paths to search for other executables. All the required executables
should be in the PATH anyway and this will fix prefixed systems.

Also add missing dependency on wget. It's in @system but it's one of
the less obvious entries.

Package-Manager: portage-2.2.28