app-editors/atom: Switch to using dev-util/electron, recompile binaries and unbundle libraries
Proper build of dev-util/electron is now in the tree (#579116),
which makes it possible to rebuild Atom properly, i.e. without
the use of precompiled blobs and with proper library unbundling.

The strategy used in this build is to take the upstream release
build and recompile the native module portions against the system
Electron and system libraries.

Gentoo-Bug: https://bugs.gentoo.org/show_bug.cgi?id=582462

Package-Manager: portage-2.2.28