python-utils-r1.eclass: stop respecting DESTTREE in python_do*/new*
Replace the code respecting current value of ${DESTTREE} in favor of
using /usr unconditionally. The feature was extremely unlikely to be
ever used (only if you did 'into ...; python_do*', quite confusing
and relied on semi-internal package manager variables.

Furthermore, it was potentially broken as it relied on DESTTREE having
an initial '/usr' value which is not warranted by the PMS.