app-antivirus/clamav-unofficial-sigs: replace v5.0.6 with v5.3.1.
A few ebuild changes were necessary. First, the man page and logrotate
scripts have gone away and have been replaced with a hokey procedure
to generate them (that we can't use in the ebuild). To keep things
working, we now ship pre-generated copies of those two files. We also
install the user.conf file which is a no-op out-of-the-box. My own
changes to os.gentoo.conf upstream now allow us to use the default
src_prepare().

Thanks are due to Tobias Sager who reported the new version and
supplied the necessary fixes.

Gentoo-Bug: 582874
Reported-By: Tobias Sager

Package-Manager: portage-2.2.26