dev-ruby/json: fix USE=doc during bootstrap
Add explicit dependencies on dev-ruby/rdoc for USE=doc.
Normally these are satified by dev-lang/ruby, but this
happens in PDEPEND. During an initial merge of ruby the
order of dev-ruby/rdoc and dev-ruby/json is not defined
so the emerging json may fail due to a missing rdoc.

Fixes bug 572042

Package-Manager: portage-2.2.28