app-admin/setools: version bump to 4.0.1, fix deps
This bumps SETools to version 4.0.1. The main change upstream is a
shift in the used icons due to a potential license incompatibility.

We also remove the dependency on qtchooser (see bug 583310) as this
is little more than a target manager (similar to eselect). The PyQt5
dependency handles the necessary dependencies already.

Finally, the libsepol dependency is also bumped as this tool requires
libsepol-2.5.

Package-Manager: portage-2.2.28