net-analyzer/nagios-check_rbl: new revision for perl-5.24 compatibility.
The tarball for this package was shipped with a copy of inc/version.pm
that was causing problems with perl-5.24. Since v1.3.7 was added to
the tree, upstream has moved SRC_URI to github, and re-released v1.3.7
with modified source code. The new tarball does not exhibit the
version.pm problem.

To fix the issue, this revision uses the new tarball, but renames it
to include the -r1 suffix so that users don't see checksum
failures. We are also removing inc/version.pm in
src_prepare(). Comments have been added to the ebuild so that we can
check for an upstream fix at a later time. The new revision also uses
EAPI=6.

Gentoo-Bug: 583966

Package-Manager: portage-2.2.28