app-admin/systemrescuecd-x86: Support running isohybrid on images
Add isohybrid flag that uses isohybrid (from syslinux) to make
the images dual-bootable as CD and USB media. This is the new
upstream-suggested method of using SRCD on USB sticks.

Since isohybrid modifies images in place and takes less than a second,
it seems reasonable to run it while installing rather than expecting
the user to copy it to a temporary location, convert it, then copy to
the media.