distutils-r1.eclass: Move creation of .pydistutils.cfg to esetup.py
Also remove .pydistutils.cfg from HOME before returning.

This config should only be needed when we are invoking setup.py. The
esetup.py wrapper provides a normalized way for doing this; if ebuilds
call setup.py without the wrapper, they are now also responsible for
setting the build directory if necessary.

This change allows us to stop overriding the HOME variable, and prevents
leakage of eclass-specifc configuration to the general ebuild
environment.