app-text/htmldoc: respect system toolchain settings
If CC and CXX variables are not exported to configure, it ignores
system toolchain setup and forces its own compiler preferences,
where clang/clang++ are on the first place if found. This is likely
not what users want and causes configure failure if CFLAGS/CXXFLAGS
contain flag(s) not supported by clang, but supported by gcc.

Package-Manager: portage-2.3.0_rc1
Signed-off-by: Andrew Savchenko <bircoph@gentoo.org>