sys-cluster/torque: dependency cleanup
Dependency atoms listed in the ebuild were rather significantly mis-matching what
is actually linked to.  Cleaned them up, added slot and slot-operators as needed.

Added libressl to IUSE but didn't test if libressl is a drop-in replacement for
openssl -- the situation should be better either way since it was a magic dep
until now.

Also bumped EAPI for torque-2.x so it is no longer deprecated, and adjusted how
the cpusets flag is added to configure.

Package-Manager: portage-2.2.28