dev-lang/php: new revision of the 7.x series to fix our webp support.
In the 5.x series, libvpx was used to provide webp support to the gd
library. It was pulled in with USE=vpx. In php-7.x, however, libvpx
has been replaced with libwebp. The old USE=vpx flag no longer worked,
so it has been replaced with USE=webp and eveything updated to use
that for php-7.x.

Package-Manager: portage-2.2.28