app-doc/doxygen: remove LINGUAS use expand, bug #584782
doxygen has language support in form of quite a number of "translator"
units. We had the habit of fine-tuning the translator units that are
actually built via the LINGUAS use expand mechanism.

But this is a broken approach due to the following reasons:

 - "translator" units are lightweight header files that are built into
   doxygen during compilation. No external dependencies are required.

 - Traditionally, the LINGUAS variable is used for packages to determine
   runtime language support (especially for GUI programs, etc) that usually
   requires additional packages to be installed - i.e., LINGUAS controls
   the languages supported in CLI and GUI applications.

   doxygen's support for translator units does *not* fit into this scheme
   at all.

 - Not all translator units have a LINGUAS equivalent (!!)

 - Hand selecting supported LINGUAS is a maintenance burden.

Therefore, I just got rid of it and install *all supported* "translator"
unconditionally.

Package-Manager: portage-2.2.28