app-crypt/gnupg: Change minimum libgcrypt version to 1.7.1
Change minimum libgcrypt version to 1.7.1. This servers at least two
purposes; (i) it ensures gnupg 2.1 is build with curve25519 support
for Gentoo users; and (ii) threads use flag is removed in this
version (as it is required to always be on), and due to (i) this makes
more sense than specifying a default interprentation for missing use
flag.

Package-Manager: portage-2.3.0_rc1