kde5.eclass: only export pkg_pretend if KDE_GCC_MINIMAL is defined
Exporting pkg_pretend is expensive, even if the exported function is a noop.

As kde5_pkg_pretend is currently only used to check if the active GCC version
is recent enough, only export that phase if it will have something to do.

The number of situations in which KDE_GCC_MINIMAL is exported will be
substantially reduced in a future commit. This will result in a large time
saving at the beginning of the emerge process as we will no longer see hundreds
of lines of:

>>> Running pre-merge checks for kde-foo/bar-123
>>> Running pre-merge checks for kde-foo/baz-123