app-shells/bash: bashrc: back out clearing of default LS_COLORS #583814
The behavior of ls in the default case (when LS_COLORS isn't set) isn't
documented well.  The manual leads you to believe the defaults will be
used when in reality they are not.  A scan of the source shows this.  So
back out some the attempts to optimize the env and go back to exporting
LS_COLORS all the time.  We'll just have to live with incompat warnings
when coreutils upgrades & changes behavior.