Revert "app-misc/wyrd: Remove := slot op from || dep, #586304"
This reverts commit 09756db2fb6945d1b767b68f4017e6f1d4e6b66a.

This commit is wrong in several aspects:
  - Such changes should come with a revbump
  - Removing := deps causes packages not to be rebuilt on upgrades, causing
    barely understandable build failures when upgrading, later on.
    Thanks to this commit, people that have installed those packages will
    likely experience those failures with the upcoming ocaml 4.03 unmask.
  - This is a non maintainer commit where no discussion at all happened with the
    maintainers during the 3 days between bug report and mass commit.
    Such discussion could have prevented the above mentioned breakage.

Proper fix is to drop the backward compatibility part of the || () dep. This fix
was slowly moving forward whenever possible and noticed. Main blocker for this is
bug #463018 which requires taking extra care for not breaking the stable tree
nor needlessly preventing future stabilization of packages.