Revert "dev-ml/ocamlnet: Remove another := slot op from || dep"
This reverts commit 3b1a4cc0f9b9a0a2c70d6bb53cb71bf059e2920e.

This commit is wrong in several aspects:
  - Such changes should come with a revbump
  - Removing := deps causes packages not to be rebuilt on upgrades, causing
    barely understandable build failures when upgrading, later on.
    Thanks to this commit, people that have installed those packages will
    likely experience those failures with the upcoming ocaml 4.03 unmask.
  - This is a non maintainer commit where no discussion at all happened with the
    maintainers during the 3 days between bug report and mass commit.
    Such discussion could have prevented the above mentioned breakage.

Proper fix is to drop the backward compatibility part of the || () dep. This fix
was slowly moving forward whenever possible and noticed. Main blocker for this is
bug #463018 which requires taking extra care for not breaking the stable tree
nor needlessly preventing future stabilization of packages.