sec-policy: Enable arm, arm64 and mips keywords inside the -9999 ebuilds
The -9999 ebuilds do not have a set KEYWORDS, but the same ebuilds are used
when creating a new release. For this purpose, the -9999 ebuilds have a
conditional code like the following:

if [[ ${PV} != 9999* ]] ; then
  KEYWORDS="~amd64 ~x86"
fi

In this commit, we update that part to include ~arm, ~arm64 and ~mips.

Package-Manager: portage-2.2.28