app-admin/needrestart: add ebuilds for version 2.8 and 9999
needrestart checks which daemons need to be restarted after library upgrades. It is inspired by checkrestart from the debian-goodies package.

needrestart supports but does not require systemd. If systemd is not available or does not return a service name needrestart uses hooks to identify the corresponding System V init script. If a process
inside a container (LXC, docker) it suggests to restart the entire container. needrestart contains a heuristic for different interpreter languages: it tries to detect if outdated script code/modules are
in use (Java, Perl, Python, Ruby).

Gentoo-bug: 588010