eclass/moz*: refactor some configuration options from mozcoreconf-v4 to mozconfig
mozilla-48 and newer have dropped some configuration options that are no longer used
in the build system, and due to the way the new python configure operates, any options
specified that are not recognized are fatal.  Moving these options out of mozcoreconf
allows the older mozconfig's to retain the same configuration while also allowing newer
mozconfig's to drop them as appropriate.