php-ext-source-r3.eclass: new revision supporting EAPI=6.
This is a new revision of the php-ext-source eclass that supports
EAPI=6 (only) and cleans up some of the existing code. The list of
user-facing changes is,

  * Support only EAPI=6.

  * PATCHES array/variable support.

  * DOCS array support (bug 512184).

  * Renamed my_conf and PHPSAPILIST variables.

Some refactoring was done, but not in a way that consumers should
notice. A migration guide can be found on the wiki:

https://wiki.gentoo.org/wiki/Project:PHP/Php-ext-source-r3_migration_guide

In addition, a number of improvements were suggested by Michał Górny
on the gentoo-dev mailing list:

https://archives.gentoo.org/gentoo-dev/message/9159cba2eabceb96d8ae16ced6631f1e

Gentoo-Bug: 512184