php-ext-pecl-r3.eclass: new revision supporting EAPI=6.
The php-ext-pecl eclasses are based mainly on the php-ext-source
eclasses. Now that we have a new revision php-ext-source-r3.eclass,
this new revision of php-ext-pecl inherits that. As a result, all of
the changes affecting that revision also affect this one. A migration
guide for users can be found on the wiki:

https://wiki.gentoo.org/wiki/Project:PHP/Php-ext-source-r3_migration_guide

Gentoo-Bug: 512184