dev-lang/php: new edit/revision to hide some elog messages when possible.
Two of the elog messages that we show only need to be dealt with
once. If the user has the recently-emerged slot of PHP in his
PHP_TARGETS already, then we don't need to bug him about that. And if
he has PHP_INI_VERSION set to a legal value, then we don't need to
inform him about that either.

A new revision is made for the php-5.6.x series. The 7.x series is not
stable, so an in-place edit was made.

Gentoo-Bug: 440574

Package-Manager: portage-2.2.28