dev-lang/php: new revisions fixing opcache installation location logic.
We install the opcache extension with a call to dolib.so, which
requires an earlier call to "into" to set its installation
location. If the user enables any SAPIs other than apache2, that
happens incidentally. However, if only the apache2 SAPI is enabled,
the installation location is never set, ultimately causing the
extension to fail to load. This is fixed by adding a call to "into"
directly before the dolib.so call.

Two calls to use_if_iuse were also replaced with the simpler "use".

Gentoo-Bug: 550676

Package-Manager: portage-2.2.28