dev-php/PEAR-HTTP_Header: new EAPI=6 revision sans the PEAR eclass.
We have a lot of trouble with our php-pear-* eclasses. Many of the
PEAR packages (including PEAR-HTTP_Header) can be installed in only a
few lines, in which case it seems smarter to skip the eclass
altogether. This new revision does so, and lets us jump to EAPI=6.

Gentoo-Bug: 576432

Package-Manager: portage-2.2.28