app-editors/neovim: Readjust dependencies
luajit is no longer a hard dependency, so we can depend on either
dev-lang/luajit, or dev-lang/lua

the perl support was erroneous, given that there is no perl client yet.

the tui is now optional, so other ui's can be used instead.

Add a clipboard use flag so neovim can copy to the clipboard.

Update to EAPI=6.