dev-php/PEAR-XML_Util: new revision to clean up dependencies.
This package had a build-time dependency on dev-lang/php that was
intended to be a runtime dependency; now fixed. There were also
redundant dependencies on dev-php/pear and dev-php/PEAR-PEAR (which
are the same, for all intents and purposes). The dev-php/pear
dependency has been removed in favor of a single PDEPEND on
dev-php/PEAR-PEAR.

Package-Manager: portage-2.2.28