net-im/minbif: fix build with newer imlib.
Newer versions of imlib have renamed ImlibLoadError to
Imlib_Load_Error. This commit adds a minimal patch to fix the
build. It's bad mojo to modify a stable ebuild, but there are
mitigating factors here. First, the stable package simply will not
build without the patch, so it's not much good to anyone. Second, this
package is maintainer-needed, and a new revision would therefore never
get stabilized to truly fix the bug for stable users. So, I've applied
the patch to the one existing stable ebuild.

Gentoo-Bug: 559412

Package-Manager: portage-2.2.28