dev-lang/python: Remove remnants of python-updater
Remove the remaining dependencies on deprecated python-updater,
and update the postinst messages appropriately. For Python 2.7,
the message has been removed altogether since it is the oldest version
supported by PYTHON_TARGETS. For the remaining versions, just inform
about potential PYTHON_TARGETS changes necessary.

The modified versions are not revision-bumped to avoid unnecessary
rebuilds due to a minor change. The versions that are still
newest-visible (i.e. not replaced by another version or revision bump)
will be bumped when python-updater is lastrited.