app-cdr/dvd+rw-tools: new revision with EAPI=6 to fix a bug.
A new revision was in order to fix bug 440426 and only show a ulimit
warning for new installs. It made sense to update to EAPI=6 for the
new revision, and that simplified the ebuild a bit. To use the EAPI=6
PATCHES support, two patches had to be revisioned to be -p1
compatible.  Theoretically they could have been modified in-place
(since epatch autodetects the -pN level), but since that would affect
a stable ebuild, I avoided it.

Gentoo-Bug: 440426

Package-Manager: portage-2.2.28