www-apache/mod_limitipconn: new revision to fix the postinst instructions.
The pkg_postinst() for this module used to suggest setting,

  APACHE2_OPTS="... -D LIMITIPCONN -D INFO"

in your apache2 conf file. The "INFO" was a mistake; it should have
been "STATUS" (to enable ExtendedStatus, as needed by mod_limitipconn).
This commit fixes the postinst message, and adds a runtime dependency
on mod_status for apache2 (previously missing).

Gentoo-Bug: 479360

Package-Manager: portage-2.2.28