dev-lang/php: new 7.x revision to fix the build on x86.
There is an upstream bug that causes the build to fail on x86 with
gcc-4.9. There was an earlier fix for gcc-4.8, and this new revision
has a fix that also works for gcc-4.9.

And, since we're no longer inheriting depend.apache.eclass, I was
finally able to update the ebuild to EAPI=6.

A redundant earlier ebuild for php-7.0.9 was also removed.

Gentoo-Bug: 571658

Package-Manager: portage-2.2.28