net-libs/libsignon-glib: revision bump with various fixes
This introduces an introspection USE flag which eliminates an automagic
dependency.

The python USE flag requires the new introspection USE flag to be enabled,
as they are introspection bindings and build but do not run without the backing
introspection files.

This revision bump also corrects the python dependencies, prunes an unused
libtool file, and switches to EAPI 6.

Package-Manager: portage-2.3.0