net-dns/djbdns: new revision with increased dnscache recursion depth.
There is a hard limit on the number of recursive queries that dnscache
will perform. This new revision adds a patch, increasing a loop
iteration limit from 100 to 200, and thus increasing the recursion
depth. The patch was based on a similar fix for OpenWrt, suggested by
John Stile.

With the new revision, the ebuild was updated to EAPI=6. The eutils
eclass was dropped as "epatch" was replaced with "eapply"; some
patches needed new revisions to work with -p1, as eapply does. Two
other patches needed revisions to avoid fuzz from the new fix.

Gentoo-Bug: 590548
Reported-By: John Stile

Package-Manager: portage-2.2.28