dev-libs/skyutils: new revision to fix build bugs.
This new revision updates to EAPI=6, and adds a patch that enables the
build to complete with -Werror=format-security. It also adds an
"sslv3" USE flag, because there is at least one package
(app-mobilephone/smssend) which should require skyutils to be build
against OpenSSL with sslv3 support.

Gentoo-Bug: 521114
Gentoo-Bug: 588326

Package-Manager: portage-2.2.28