media-tv/dvbstream: new revision respecting LDFLAGS.
This is a new revision with an update to EAPI=6 and a little ebuild
cleanup. Some missing die() statements were added, and some file
copying was avoided by using the newbin/newdoc helpers. Most
importantly, Michael Mair-Keimberger's fix to support LDFLAGS was
incorporated.

I built the package on amd64, so I've keyworded it ~amd64 as well.

Gentoo-Bug: 337327

Package-Manager: portage-2.2.28